refactor: ペイン表示領域の計測処理を共通化
